fix: refactor button, chip, loader and checkbox [MDS-52]
