+++ packages/core/src/button/Button.tsx
@@ -1,27 +1,11 @@
 import React, { useState } from 'react';
 import Hover from '../hover/Hover';
 import Size from '../private/enums/Size';
-import AnimationContent from './private/buttonAnimations/AnimationContent';
+import AnimationContent from './styles/buttonAnimations/AnimationContent';
 import ButtonComponent from './styles/ButtonComponent';
-
-type ButtonSizes = Size.SMALL | Size.MEDIUM | Size.LARGE | Size.XLARGE;
-type ButtonVariants = 'primary' | 'secondary' | 'tertiary' | 'ghost';
-type ButtonIcons = 'left' | 'right' | 'only';
-type ButtonAnimations = 'progress' | 'success' | 'error' | 'pulse';
-
-export interface ButtonProps
-  extends React.ButtonHTMLAttributes<HTMLButtonElement> {
-  variant?: ButtonVariants;
-  size?: ButtonSizes | 'small' | 'medium' | 'large' | 'xlarge';
-  icon?: ButtonIcons;
-  animation?: ButtonAnimations;
-  iconSize?: number;
-  iconLeft?: JSX.Element | boolean;
-  iconRight?: JSX.Element | boolean;
-  iconOnly?: JSX.Element | boolean;
-  disabled?: boolean;
-  fullWidth?: boolean;
-}
+import type ButtonProps from './private/types/ButtonProps';
+import type ButtonSizesProps from './private/types/ButtonSizesProps';
+import type ButtonVariantsProps from './private/types/ButtonVariantsProps';
 
 const Button: React.FC<ButtonProps> = ({
   children,
@@ -42,8 +26,8 @@ const Button: React.FC<ButtonProps> = ({
     animation === 'progress' || animation === 'success';
   return (
     <ButtonComponent
-      size={size as ButtonSizes}
-      variant={variant as ButtonVariants}
+      size={size as ButtonSizesProps}
+      variant={variant as ButtonVariantsProps}
       icon={icon}
       iconSize={iconSize}
       iconLeft={iconLeft}
