+++ packages/core/src/button/private/types/ButtonAnimationsProps.ts
@@ -1,0 +1,3 @@
+type ButtonAnimationsProps = 'progress' | 'success' | 'error' | 'pulse';
+
+export default ButtonAnimationsProps;
