+++ packages/core/src/button/private/types/ButtonIconsProps.ts
@@ -1,0 +1,3 @@
+type ButtonIconsProps = 'left' | 'right' | 'only';
+
+export default ButtonIconsProps;
