+++ packages/core/src/button/private/types/ButtonSizesProps.ts
@@ -1,0 +1,13 @@
+import Size from '../../../private/enums/Size';
+
+type ButtonSizesProps =
+  | Size.SMALL
+  | Size.MEDIUM
+  | Size.LARGE
+  | Size.XLARGE
+  | 'small'
+  | 'medium'
+  | 'large'
+  | 'xlarge';
+
+export default ButtonSizesProps;
