+++ packages/core/src/button/private/types/ButtonVariantsProps.ts
@@ -1,0 +1,3 @@
+type ButtonVariantsProps = 'primary' | 'secondary' | 'tertiary' | 'ghost';
+
+export default ButtonVariantsProps;
