+++ packages/core/src/button/styles/buttonAnimations/errorAnimation.ts
