+++ packages/core/src/button/styles/buttonAnimations/pulseAnimation.ts
