+++ packages/core/src/button/styles/buttonSizes/largePadding.ts
@@ -1,5 +1,5 @@
 import { rem } from '@heathmont/moon-utils';
-import type { ButtonProps } from '../../Button';
+import type ButtonProps from '../../private/types/ButtonProps';
 
 const largePadding = ({ icon, iconLeft, iconRight, iconOnly }: ButtonProps) => {
   if (icon === 'left' || iconLeft) {
