--- packages/core/src/checkbox/@types/Variants.d.ts
@@ -1,1 +1,0 @@
-type Variants = 'option';
