+++ packages/core/src/checkbox/private/types/CheckboxIconWrapperProps.ts
@@ -1,0 +1,7 @@
+type CheckboxIconWrapperProps = {
+  checked?: boolean;
+  label?: string | JSX.Element;
+  dir?: 'ltr' | 'rtl' | 'auto';
+};
+
+export default CheckboxIconWrapperProps;
