+++ packages/core/src/checkbox/private/types/VariantProps.ts
@@ -1,0 +1,3 @@
+type VariantProps = 'option';
+
+export default VariantProps;
