+++ packages/core/src/chip/private/types/SizeProps.ts
@@ -1,0 +1,5 @@
+import Size from '../../../private/enums/Size';
+
+type SizeProps = Size.SMALL | Size.MEDIUM | 'small' | 'medium';
+
+export default SizeProps;
