+++ packages/core/src/chip/private/types/StyledChipProps.ts
@@ -1,0 +1,13 @@
+import type SizeProps from './SizeProps';
+
+type StyledChipProps = {
+  isActive?: boolean;
+  isStroke?: boolean;
+  size?: SizeProps;
+  iconSize?: number;
+  iconLeft?: JSX.Element | boolean;
+  iconRight?: JSX.Element | boolean;
+  iconOnly?: JSX.Element | boolean;
+};
+
+export default StyledChipProps;
