--- packages/core/src/loader/private/Size.ts
@@ -1,1 +1,0 @@
-export type Size = 'twoxsmall' | 'xsmall' | 'small' | 'medium' | 'large';
