+++ packages/core/src/loader/private/types/LoaderProps.ts
@@ -1,0 +1,9 @@
+import type SizeProps from './SizeProps';
+import type { ColorProps } from '@heathmont/moon-themes';
+
+type LoaderProps = {
+  color?: ColorProps;
+  size?: SizeProps;
+};
+
+export default LoaderProps;
