+++ packages/core/src/loader/private/types/SizeProps.ts
@@ -1,0 +1,15 @@
+import Size from '../../../private/enums/Size';
+
+type SizeProps =
+  | Size.TWOXSMALL
+  | Size.XSMALL
+  | Size.SMALL
+  | Size.MEDIUM
+  | Size.LARGE
+  | 'twoxsmall'
+  | 'xsmall'
+  | 'small'
+  | 'medium'
+  | 'large';
+
+export default SizeProps;
