+++ packages/core/src/loader/private/utils/setColor.ts
